services/decision-simulation/requirements.txt
fastapi==0.115.4
uvicorn==0.34.0
google-generativeai==0.8.3
chromadb==0.4.24
sentence-transformers==3.3.1
langchain==0.1.20
simpy==4.1.1
pydantic==2.9.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
boto3==1.35.0
httpx==0.27.2

